+++ model_navigator/pipelines/builders/tensorflow_tensorrt.py
@@ -18,12 +18,8 @@ from typing import Dict, List
 from model_navigator.api.config import DeviceKind, Format
 from model_navigator.commands.base import ExecutionUnit
 from model_navigator.commands.convert.tf import ConvertSavedModel2TFTRT
-from model_navigator.commands.delete_model import DeleteModel
-from model_navigator.commands.performance.performance import Performance
-from model_navigator.commands.tensorrt_profile_builder import TensorRTProfileBuilder
 from model_navigator.configuration.common_config import CommonConfig
 from model_navigator.configuration.model.model_config import ModelConfig
-from model_navigator.frameworks.tensorrt.utils import search_for_optimized_profiles
 from model_navigator.pipelines.pipeline import Pipeline
 from model_navigator.runners.registry import get_runner
 from model_navigator.runners.tensorflow import TensorFlowTensorRTRunner
@@ -45,34 +41,34 @@ def tensorflow_tensorrt_conversion_builder(
         return Pipeline(name="TensorFlow-TensorRT Conversion", execution_units=[])
 
     tensorflow_trt_models_config = models_config.get(Format.TF_TRT, [])
-    run_profiles_search = search_for_optimized_profiles(config, tensorflow_trt_models_config)
+    # run_profiles_search = search_for_optimized_profiles(config, tensorflow_trt_models_config)
 
     execution_units: List[ExecutionUnit] = []
     for model_cfg in tensorflow_trt_models_config:
-        if run_profiles_search:
-            # Run initial conversion to TensorFlow TensorRT
-            execution_units.append(ExecutionUnit(command=ConvertSavedModel2TFTRT, model_config=model_cfg))
-
-            # Generate preliminary profiling results
-            execution_units.append(
-                ExecutionUnit(
-                    command=Performance,
-                    model_config=model_cfg,
-                    runner_cls=get_runner(TensorFlowTensorRTRunner),
-                )
-            )
-
-            # Delete temporary TensorFlow TensorRT models
-            execution_units.append(ExecutionUnit(command=DeleteModel, model_config=model_cfg))
-
-        # Generate TensorRT profiles or use user provided ones
-        execution_units.append(
-            ExecutionUnit(
-                command=TensorRTProfileBuilder,
-                model_config=model_cfg,
-                results_lookup_runner_cls=get_runner(TensorFlowTensorRTRunner),
-            )
-        )
+        # if run_profiles_search:
+        #     # Run initial conversion to TensorFlow TensorRT
+        #     execution_units.append(ExecutionUnit(command=ConvertSavedModel2TFTRT, model_config=model_cfg))
+        #
+        #     # Generate preliminary profiling results
+        #     execution_units.append(
+        #         ExecutionUnit(
+        #             command=Performance,
+        #             model_config=model_cfg,
+        #             runner_cls=get_runner(TensorFlowTensorRTRunner),
+        #         )
+        #     )
+        #
+        #     # Delete temporary TensorFlow TensorRT models
+        #     execution_units.append(ExecutionUnit(command=DeleteModel, model_config=model_cfg))
+        #
+        # # Generate TensorRT profiles or use user provided ones
+        # execution_units.append(
+        #     ExecutionUnit(
+        #         command=TensorRTProfileBuilder,
+        #         model_config=model_cfg,
+        #         results_lookup_runner_cls=get_runner(TensorFlowTensorRTRunner),
+        #     )
+        # )
 
         # Convert to TensorFlow TensorRT again, this time with optimized profiles
         execution_units.append(