+++ model_navigator/pipelines/builders/tensorrt.py
@@ -17,13 +17,9 @@ from typing import Dict, List
 
 from model_navigator.api.config import DeviceKind, Format
 from model_navigator.commands.base import ExecutionUnit
-from model_navigator.commands.delete_model import DeleteModel
-from model_navigator.commands.performance.performance import Performance
-from model_navigator.commands.tensorrt_profile_builder import TensorRTProfileBuilder
 from model_navigator.configuration.common_config import CommonConfig
 from model_navigator.configuration.model.model_config import ModelConfig
 from model_navigator.frameworks import is_trt_available
-from model_navigator.frameworks.tensorrt.utils import search_for_optimized_profiles
 from model_navigator.pipelines.pipeline import Pipeline
 from model_navigator.runners.registry import get_runner
 from model_navigator.runners.tensorrt import TensorRTRunner
@@ -43,36 +39,36 @@ def tensorrt_conversion_builder(config: CommonConfig, models_config: Dict[Format
         return Pipeline(name="TensorRT Conversion", execution_units=[])
 
     trt_models_config = models_config.get(Format.TENSORRT, [])
-    run_profiles_search = search_for_optimized_profiles(config, trt_models_config)
+    # run_profiles_search = search_for_optimized_profiles(config, trt_models_config)
 
     from model_navigator.commands.convert.onnx.onnx2trt import ConvertONNX2TRT
 
     execution_units: List[ExecutionUnit] = []
     for model_cfg in trt_models_config:
-        if run_profiles_search:
-            # Run initial conversion to TensorRT
-            execution_units.append(ExecutionUnit(command=ConvertONNX2TRT, model_config=model_cfg))
-
-            # Generate preliminary profiling results
-            execution_units.append(
-                ExecutionUnit(
-                    command=Performance,
-                    model_config=model_cfg,
-                    runner_cls=get_runner(TensorRTRunner),
-                )
-            )
-
-            # Delete temporary TensorRT models
-            execution_units.append(ExecutionUnit(command=DeleteModel, model_config=model_cfg))
+        # if run_profiles_search:
+        #     # Run initial conversion to TensorRT
+        #     execution_units.append(ExecutionUnit(command=ConvertONNX2TRT, model_config=model_cfg))
+        #
+        #     # Generate preliminary profiling results
+        #     execution_units.append(
+        #         ExecutionUnit(
+        #             command=Performance,
+        #             model_config=model_cfg,
+        #             runner_cls=get_runner(TensorRTRunner),
+        #         )
+        #     )
+        #
+        #     # Delete temporary TensorRT models
+        #     execution_units.append(ExecutionUnit(command=DeleteModel, model_config=model_cfg))
 
         # Generate TensorRT profiles or use user provided ones
-        execution_units.append(
-            ExecutionUnit(
-                command=TensorRTProfileBuilder,
-                model_config=model_cfg,
-                results_lookup_runner_cls=get_runner(TensorRTRunner),
-            )
-        )
+        # execution_units.append(
+        #     ExecutionUnit(
+        #         command=TensorRTProfileBuilder,
+        #         model_config=model_cfg,
+        #         results_lookup_runner_cls=get_runner(TensorRTRunner),
+        #     )
+        # )
 
         # Convert ONNX to TensorRT again, this time with optimized profiles
         execution_units.append(