+++ model_navigator/pipelines/builders/torch_tensorrt.py
@@ -18,12 +18,8 @@ from typing import Dict, List
 from model_navigator.api.config import DeviceKind, Format
 from model_navigator.commands.base import ExecutionUnit
 from model_navigator.commands.convert.torch import ConvertTorchScript2TorchTensorRT
-from model_navigator.commands.delete_model import DeleteModel
-from model_navigator.commands.performance.performance import Performance
-from model_navigator.commands.tensorrt_profile_builder import TensorRTProfileBuilder
 from model_navigator.configuration.common_config import CommonConfig
 from model_navigator.configuration.model.model_config import ModelConfig
-from model_navigator.frameworks.tensorrt.utils import search_for_optimized_profiles
 from model_navigator.pipelines.pipeline import Pipeline
 from model_navigator.runners.registry import get_runner
 from model_navigator.runners.torch import TorchTensorRTRunner
@@ -43,34 +39,34 @@ def torch_tensorrt_conversion_builder(config: CommonConfig, models_config: Dict[
         return Pipeline(name="Torch-TensorRT Conversion", execution_units=[])
 
     torch_trt_models_config = models_config.get(Format.TORCH_TRT, [])
-    run_profiles_search = search_for_optimized_profiles(config, torch_trt_models_config)
+    # run_profiles_search = search_for_optimized_profiles(config, torch_trt_models_config)
 
     execution_units: List[ExecutionUnit] = []
     for model_cfg in torch_trt_models_config:
-        if run_profiles_search:
-            # Run initial conversion to Torch TensorRT
-            execution_units.append(ExecutionUnit(command=ConvertTorchScript2TorchTensorRT, model_config=model_cfg))
-
-            # Generate preliminary profiling results
-            execution_units.append(
-                ExecutionUnit(
-                    command=Performance,
-                    model_config=model_cfg,
-                    runner_cls=get_runner(TorchTensorRTRunner),
-                )
-            )
-
-            # Delete temporary Torch TensorRT models
-            execution_units.append(ExecutionUnit(command=DeleteModel, model_config=model_cfg))
+        # if run_profiles_search:
+        #     # Run initial conversion to Torch TensorRT
+        #     execution_units.append(ExecutionUnit(command=ConvertTorchScript2TorchTensorRT, model_config=model_cfg))
+        #
+        #     # Generate preliminary profiling results
+        #     execution_units.append(
+        #         ExecutionUnit(
+        #             command=Performance,
+        #             model_config=model_cfg,
+        #             runner_cls=get_runner(TorchTensorRTRunner),
+        #         )
+        #     )
+        #
+        #     # Delete temporary Torch TensorRT models
+        #     execution_units.append(ExecutionUnit(command=DeleteModel, model_config=model_cfg))
 
         # Generate TensorRT profiles or use user provided ones
-        execution_units.append(
-            ExecutionUnit(
-                command=TensorRTProfileBuilder,
-                model_config=model_cfg,
-                results_lookup_runner_cls=get_runner(TorchTensorRTRunner),
-            )
-        )
+        # execution_units.append(
+        #     ExecutionUnit(
+        #         command=TensorRTProfileBuilder,
+        #         model_config=model_cfg,
+        #         results_lookup_runner_cls=get_runner(TorchTensorRTRunner),
+        #     )
+        # )
 
         # Convert TorchScript to Torch TensorRT again, this time with optimized profiles
         execution_units.append(